containers/app/requirements.txt
Flask
requests
redis
docker
confluent_kafka
kubernetes
grpcio-tools
grpcio
aiohttp
quart
